+++ test/helpers/helpers.dart
@@ -1,0 +1,1 @@
+Future get tick => Future.microtask(() {});